refactor(sidebar): extract SidebarWorktreeRow, StaticWorktreeRow, useScrollIndicator, and useRecipeDialogState from SidebarContent
- Extract SidebarWorktreeRow (~170 lines, React.memo sortable row) to SidebarWorktreeRow.tsx
- Extract StaticWorktreeRow (~85 lines, React.memo non-sortable row) to StaticWorktreeRow.tsx
- Extract useScrollIndicator hook (~80 lines) to useScrollIndicator.ts
- Extract useRecipeDialogState hook (~90 lines) to useRecipeDialogState.ts
- Reduce SidebarContent.tsx from 1356 to 925 lines
- Update grid.test.ts to read StaticWorktreeRow from new file path
- Preserve React.memo wrappers, internal useWorktreeStore subscriptions, and AbortController cleanup patterns
