+++ src/components/Sidebar/StaticWorktreeRow.tsx
@@ -1,0 +1,100 @@
+import React, { useCallback, useMemo } from "react";
+import { ErrorBoundary } from "@/components/ErrorBoundary";
+import { WorktreeCard, type WorktreeCardProps } from "@/components/Worktree";
+import { WorktreeCardErrorFallback } from "@/components/Worktree/WorktreeCardErrorFallback";
+import { useWorktreeStore } from "@/hooks/useWorktreeStore";
+import type { WorktreeState } from "@/types";
+import type { WorktreeActions } from "@/hooks/useWorktreeActions";
+import type { UseAgentLauncherReturn } from "@/hooks/useAgentLauncher";
+
+interface StaticWorktreeRowProps {
+  worktreeId: string;
+  activeWorktreeId: string | null;
+  focusedWorktreeId: string | null;
+  totalWorktreeCount: number;
+  selectWorktree: (id: string) => void;
+  worktreeActions: WorktreeActions;
+  availability: UseAgentLauncherReturn["availability"];
+  agentSettings: UseAgentLauncherReturn["agentSettings"];
+  homeDir: string | undefined;
+  aggregateCounts?: WorktreeCardProps["aggregateCounts"];
+}
+
+const StaticWorktreeRow = React.memo(function StaticWorktreeRow({
+  worktreeId,
+  activeWorktreeId,
+  focusedWorktreeId,
+  totalWorktreeCount,
+  selectWorktree,
+  worktreeActions,
+  availability,
+  agentSettings,
+  homeDir,
+  aggregateCounts,
+}: StaticWorktreeRowProps) {
+  const worktreeSnap = useWorktreeStore((state) => state.worktrees.get(worktreeId));
+  const worktree = useMemo(
+    () =>
+      worktreeSnap
+        ? ({
+            ...worktreeSnap,
+            worktreeChanges: worktreeSnap.worktreeChanges ?? null,
+            lastActivityTimestamp: worktreeSnap.lastActivityTimestamp ?? null,
+          } as WorktreeState)
+        : undefined,
+    [worktreeSnap]
+  );
+
+  const onSelect = useCallback(() => selectWorktree(worktreeId), [selectWorktree, worktreeId]);
+  const onCopyTree = useCallback(
+    () => worktree && worktreeActions.handleCopyTree(worktree),
+    [worktree, worktreeActions]
+  );
+  const onOpenEditor = useCallback(
+    () => worktree && worktreeActions.handleOpenEditor(worktree),
+    [worktree, worktreeActions]
+  );
+  const onSaveLayout = useCallback(
+    () => worktree && worktreeActions.handleSaveLayout(worktree),
+    [worktree, worktreeActions]
+  );
+  const onLaunchAgent = useCallback(
+    (agentId: string) => worktreeActions.handleLaunchAgent(worktreeId, agentId),
+    [worktreeActions, worktreeId]
+  );
+
+  if (!worktree) return null;
+
+  return (
+    <div role="row" data-worktree-row={worktreeId} tabIndex={-1}>
+      <div role="gridcell">
+        <ErrorBoundary
+          variant="component"
+          componentName="WorktreeCard"
+          fallback={WorktreeCardErrorFallback}
+          resetKeys={[worktreeId]}
+          context={{ worktreeId }}
+        >
+          <WorktreeCard
+            worktree={worktree}
+            isActive={worktreeId === activeWorktreeId}
+            isFocused={worktreeId === focusedWorktreeId}
+            isSingleWorktree={totalWorktreeCount === 1}
+            aggregateCounts={aggregateCounts}
+            onSelect={onSelect}
+            onCopyTree={onCopyTree}
+            onOpenEditor={onOpenEditor}
+            onSaveLayout={onSaveLayout}
+            onLaunchAgent={onLaunchAgent}
+            agentAvailability={availability}
+            agentSettings={agentSettings}
+            homeDir={homeDir}
+          />
+        </ErrorBoundary>
+      </div>
+    </div>
+  );
+});
+
+export { StaticWorktreeRow };
+export type { StaticWorktreeRowProps };
